dynamic input works
