+++ components/forms/add-recipe-form.tsx
@@ -1,9 +1,10 @@
 import { View, Text, StyleSheet, Pressable, TextInput } from "react-native";
-import {Picker} from '@react-native-picker/picker';
-import React, { useState } from 'react';
+import { Picker } from "@react-native-picker/picker";
+import React, { useState } from "react";
 // import { TextInput } from "react-native-gesture-handler";
 import { Ingredient, Recipe } from "../../types/recipe.type";
 import IngredientInput from "../inputs/ingredient";
+import FormPart from "../inputs/ingredient";
 
 const AddRecipeForm = () => {
   const [recipe, setRecipe] = useState<Recipe>();
@@ -15,47 +16,24 @@ const AddRecipeForm = () => {
   // // our number of ingredient inputs
   // const [numInputs, setNumInputs] = useState(1);
 
-  const FormPart = ({ index, data, onChange, onRemove }) => {
-    return (
-      <View key={index}>
-        <TextInput
-          value={data.name}
-          onChangeText={(text) => onChange(index, { ...data, name: text })}
-          placeholder="Name"
-        />
-        <TextInput
-          value={data.description}
-          onChangeText={(text) => onChange(index, { ...data, description: text })}
-          placeholder="Description"
-        />
-        <Picker
-          selectedValue={data.option}
-          onValueChange={(value) => onChange(index, { ...data, option: value })}
-        >
-          <Picker.Item label="Option 1" value="option1" />
-          <Picker.Item label="Option 2" value="option2" />
-          {/* Add more Picker.Item elements for additional options */}
-        </Picker>
-        {index > 0 && <Pressable onPress={() => onRemove(index)}>Remove</Pressable>}
-      </View>
+  const [formData, setFormData] = useState([]);
+
+  const handleAddPart = () => {
+    setFormData([
+      ...formData,
+      { name: "", description: "", option: "option1" },
+    ]);
+  };
+
+  const handleChange = (index, newData) => {
+    setFormData((prevData) =>
+      prevData.map((item, i) => (i === index ? newData : item))
     );
   };
 
-    const [formData, setFormData] = useState([]);
-  
-    const handleAddPart = () => {
-      setFormData([...formData, { name: '', description: '', option: 'option1' }]);
-    };
-  
-    const handleChange = (index, newData) => {
-      setFormData((prevData) =>
-        prevData.map((item, i) => (i === index ? newData : item))
-      );
-    };
-  
-    const handleRemovePart = (index) => {
-      setFormData((prevData) => prevData.filter((_, i) => i !== index));
-    };
+  const handleRemovePart = (index) => {
+    setFormData((prevData) => prevData.filter((_, i) => i !== index));
+  };
 
   return (
     <View>
@@ -66,21 +44,23 @@ const AddRecipeForm = () => {
         placeholder="Recipe Name"
       /> */}
       <View>
-      {formData.map((data, index) => (
-        <FormPart
-          key={index}
-          index={index}
-          data={data}
-          onChange={handleChange}
-          onRemove={handleRemovePart}
-        />
-      ))}
-      <Pressable onPress={handleAddPart}>Add Another Part</Pressable>
-    </View>
+        {formData.map((data, index) => (
+          <FormPart
+            key={index}
+            index={index}
+            data={data}
+            onChange={handleChange}
+            onRemove={handleRemovePart}
+          />
+        ))}
+        <Pressable onPress={handleAddPart}>
+          <Text>Add Another Part</Text>
+        </Pressable>
+      </View>
       {/* Test button */}
-      {/* <Pressable onPress={() => console.log(ingredients)}>
+      <Pressable onPress={() => console.log(formData)}>
         <Text>TEST</Text>
-      </Pressable> */}
+      </Pressable>
     </View>
   );
 };
